refactor: add `ArgillaDatasetMixin`
